backend/requirements.txt
# FastAPI server
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo[srv]==4.7.3

# LangChain core (0.3)
langchain==0.3.7
langchain-community==0.3.7

# MongoDB integration
langchain-mongodb==0.2.0

# Google Gemini
langchain-google-genai==2.0.2
google-generativeai==0.8.3

# Environment variables
python-dotenv==1.0.1

# Async HTTP & data validation
httpx==0.27.0
pydantic==2.7.4

# Retries / robustness
tenacity==8.3.0
langfuse==2.53.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
